After the work on the networked chat server for Samu is done, now I have started to work on a network chat client.
The two typical use cases may be the following:

- a Samu program talks to another Samu program
- a Samu program talks to himself

What I'd really like, of course, is not to have to code this as follows:

main()
{
...
if(random() < a_magic_number)
 Samu_starts_talk();
...
}

but he should do it of his own free will :) A similar suggestion has been already made in paper http://arxiv.org/abs/1511.02889:

"This is also reinforced by the fact that the incremental learning was implemented in the main.cpp and not in a class of Samu. From a point of view of a programmer, it means that the incremental approach is not yet really part of Samu."

In the case of starting conversations, it means Samu's free will must be based on his learning algorithm. We are going to focus on trying to find a solution to this challenge in the next edition of Samu.